// apps/console/src/__generated__/core/RiskControlsPage_risk.graphql.ts
export type RiskControlsPage_risk$data = {
  readonly controls: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly framework: {
          readonly id: string;
          readonly name: string;
        };
        readonly id: string;
        readonly name: string;
        readonly sectionTitle: string;
      };
    }>;
  };
  readonly id: string;
  readonly " $fragmentType": "RiskControlsPage_risk";
};
export type RiskControlsPage_risk$key = {
  readonly " $data"?: RiskControlsPage_risk$data;
  readonly " $fragmentSpreads": FragmentRefs<"RiskControlsPage_risk">;
};